Move Rugged errors to Exceptions class
